Simplify and consolidate test suite
- Focus container tests exclusively on Podman
- Simplify testing approach with clear file organization
- Remove unnecessary test files
- Update CI workflow to use new test file names
- Update tests README with simplified approach
- Remove deprecation markers and cleanup code
