--- tests/e2e/test_container.py
@@ -1,546 +1,0 @@
-"""
-End-to-end test for MCP server container.
-This test:
-1. Ensures Podman is available
-2. Ensures the image is built
-3. Tests running the container with simple commands
-4. Tests MCP server functionality
-"""
-
-import subprocess
-import os
-import sys
-import pytest
-from pathlib import Path
-
-# Get the project root directory
-PROJECT_ROOT = Path(__file__).parents[2].absolute()
-
-# Mark all tests in this file
-pytestmark = [pytest.mark.e2e, pytest.mark.container]
-
-
-def test_containerfile_exists():
-    """Test that the Containerfile exists in the project directory."""
-    containerfile_path = PROJECT_ROOT / "Containerfile"
-    assert containerfile_path.exists(), "Containerfile does not exist"
-
-
-def test_container_build():
-    """Test that the container image builds successfully."""
-    containerfile_path = PROJECT_ROOT / "Containerfile"
-
-    # Check Containerfile exists
-    assert containerfile_path.exists(), "Containerfile does not exist"
-
-    # Check that Podman is available
-    try:
-        subprocess.run(
-            ["podman", "--version"],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-            timeout=5,
-        )
-    except (subprocess.SubprocessError, FileNotFoundError, subprocess.TimeoutExpired):
-        pytest.skip("Podman is not available")
-
-    # Use a unique tag for testing
-    test_tag = "mcp-server-troubleshoot:test-build"
-
-    try:
-        # Build the image
-        result = subprocess.run(
-            ["podman", "build", "-t", test_tag, "-f", "Containerfile", "."],
-            cwd=str(PROJECT_ROOT),
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            check=True,
-            timeout=300,  # 5 minutes timeout for build
-        )
-
-        # Check if build succeeded
-        assert result.returncode == 0, f"Container build failed: {result.stderr}"
-
-        # Verify image exists
-        image_check = subprocess.run(
-            ["podman", "image", "exists", test_tag],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-            check=False,
-        )
-        assert image_check.returncode == 0, f"Image {test_tag} not found after build"
-
-    except subprocess.CalledProcessError as e:
-        pytest.fail(f"Container build failed with error: {e.stderr}")
-
-    finally:
-        # Clean up the test image
-        subprocess.run(
-            ["podman", "rmi", "-f", test_tag],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-            check=False,
-        )
-
-
-def cleanup_test_container():
-    """Remove any existing test container."""
-    subprocess.run(
-        ["podman", "rm", "-f", "mcp-test"], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL
-    )
-
-
-@pytest.fixture
-def container_setup(docker_image, ensure_bundles_directory):
-    """Setup Podman environment for testing."""
-    # The docker_image fixture ensures Podman is available and the image is built
-    # The ensure_bundles_directory fixture creates and returns the bundles directory
-
-    # Get bundles directory
-    bundles_dir = ensure_bundles_directory
-
-    # Clean up any existing test container
-    cleanup_test_container()
-
-    # Set test token
-    os.environ["SBCTL_TOKEN"] = "test-token"
-
-    yield bundles_dir
-
-    # Cleanup after tests
-    cleanup_test_container()
-
-
-def test_basic_container_functionality(container_setup):
-    """Test that the container can run basic commands."""
-    bundles_dir = container_setup
-
-    result = subprocess.run(
-        [
-            "podman",
-            "run",
-            "--name",
-            "mcp-test",
-            "--rm",
-            "-v",
-            f"{bundles_dir}:/data/bundles",
-            "-e",
-            f"SBCTL_TOKEN={os.environ.get('SBCTL_TOKEN', 'test-token')}",
-            "--entrypoint",
-            "/bin/bash",
-            "mcp-server-troubleshoot:latest",
-            "-c",
-            "echo 'Container is working!'",
-        ],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        text=True,
-        check=True,
-    )
-
-    assert "Container is working!" in result.stdout
-
-
-def test_python_functionality(container_setup):
-    """Test that Python works in the container."""
-    bundles_dir = container_setup
-
-    result = subprocess.run(
-        [
-            "podman",
-            "run",
-            "--name",
-            "mcp-test",
-            "--rm",
-            "-v",
-            f"{bundles_dir}:/data/bundles",
-            "-e",
-            f"SBCTL_TOKEN={os.environ.get('SBCTL_TOKEN', 'test-token')}",
-            "--entrypoint",
-            "/bin/bash",
-            "mcp-server-troubleshoot:latest",
-            "-c",
-            "python --version",
-        ],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        text=True,
-        check=True,
-    )
-
-    version_output = result.stdout.strip() or result.stderr.strip()
-    assert "Python" in version_output
-
-
-def test_mcp_cli(container_setup):
-    """Test that the MCP server CLI works in the container."""
-    bundles_dir = container_setup
-
-    result = subprocess.run(
-        [
-            "podman",
-            "run",
-            "--name",
-            "mcp-test",
-            "--rm",
-            "-v",
-            f"{bundles_dir}:/data/bundles",
-            "-e",
-            f"SBCTL_TOKEN={os.environ.get('SBCTL_TOKEN', 'test-token')}",
-            "--entrypoint",
-            "/bin/bash",
-            "mcp-server-troubleshoot:latest",
-            "-c",
-            "python -m mcp_server_troubleshoot.cli --help",
-        ],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        text=True,
-    )
-
-    combined_output = result.stdout + result.stderr
-    assert "usage:" in combined_output.lower() or result.returncode == 0
-
-
-@pytest.mark.timeout(30)  # Set a 30-second timeout for this test
-def test_mcp_protocol(container_setup, docker_image):
-    """
-    Test MCP protocol communication with the container.
-
-    This test sends a JSON-RPC request to the container running in MCP mode
-    and verifies that it responds correctly.
-    """
-    import uuid
-    import tempfile
-    import time
-    from pathlib import Path
-
-    # Create a temporary directory for the bundle
-    with tempfile.TemporaryDirectory() as temp_dir:
-        temp_path = Path(temp_dir)
-
-        # Generate a unique container ID for this test
-        container_id = f"mcp-test-{uuid.uuid4().hex[:8]}"
-
-        # Make sure there's no container with this name already
-        subprocess.run(
-            ["podman", "rm", "-f", container_id],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-            check=False,
-        )
-
-        # Start the container using run instead of Popen
-        print(f"Starting test container: {container_id}")
-
-        # Use detached mode to run in background
-        container_start = subprocess.run(
-            [
-                "podman",
-                "run",
-                "--name",
-                container_id,
-                "-d",  # Detached mode
-                "-i",  # Interactive mode for stdin
-                "-v",
-                f"{temp_path}:/data/bundles",
-                "-e",
-                "SBCTL_TOKEN=test-token",
-                "-e",
-                "MCP_BUNDLE_STORAGE=/data/bundles",
-                docker_image,
-            ],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            text=True,
-            check=False,
-        )
-
-        # Print full container start output for debugging
-        print(f"Container start stdout: {container_start.stdout}")
-        print(f"Container start stderr: {container_start.stderr}")
-        print(f"Container start return code: {container_start.returncode}")
-
-        if container_start.returncode != 0:
-            print(f"Failed to start container: {container_start.stderr}")
-            pytest.fail(f"Failed to start container: {container_start.stderr}")
-
-        try:
-            # Wait a moment for the container to start
-            time.sleep(2)
-
-            # Check if the container started successfully with detailed logging
-            ps_check = subprocess.run(
-                ["podman", "ps", "-a", "--format", "{{.ID}} {{.Names}} {{.Status}}"],
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-            )
-
-            print(f"Container status: {ps_check.stdout}")
-
-            # Also get logs in case it failed to start properly
-            logs_check = subprocess.run(
-                ["podman", "logs", container_id],
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                check=False,
-            )
-
-            print(f"Container logs stdout: {logs_check.stdout}")
-            print(f"Container logs stderr: {logs_check.stderr}")
-
-            # Check specifically for this container
-            running_check = subprocess.run(
-                ["podman", "ps", "-q", "-f", f"name={container_id}"],
-                stdout=subprocess.PIPE,
-                text=True,
-            )
-
-            assert running_check.stdout.strip(), "Podman container failed to start"
-
-            # Instead of using a full client, we'll use a simpler approach
-            # to verify basic MCP functionality
-
-            # Simple version check - we expect to get a response, even if it's an error
-            from threading import Timer
-
-            def timeout_handler():
-                print("Test timed out, terminating container...")
-                subprocess.run(
-                    ["podman", "rm", "-f", container_id],
-                    stdout=subprocess.DEVNULL,
-                    stderr=subprocess.DEVNULL,
-                    check=False,
-                    timeout=5,
-                )
-                pytest.fail("Test timed out waiting for response")
-
-            # Set a timer for timeout
-            timer = Timer(10.0, timeout_handler)
-            timer.start()
-
-            try:
-                # Wait a bit longer for container to produce logs
-                time.sleep(3)
-
-                # Instead of checking logs, let's just check the container is running
-                ps_check_detailed = subprocess.run(
-                    [
-                        "podman",
-                        "ps",
-                        "--format",
-                        "{{.Command}},{{.Status}}",
-                        "-f",
-                        f"name={container_id}",
-                    ],
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.PIPE,
-                    text=True,
-                    check=False,
-                )
-
-                # Print detailed info for debugging
-                print(f"Container info: {ps_check_detailed.stdout}")
-
-                # Just check that the container started and is running
-                assert ps_check_detailed.stdout.strip(), "Container is not running"
-                # Podman may truncate or format the command differently than Docker
-                # Just check that the container is running (we already know it's our mcp-server)
-                assert "Up" in ps_check_detailed.stdout, "Container is not in 'Up' state"
-
-                # Consider the test passed if container is running
-                print("Container is running properly")
-
-                # Skip the MCP protocol communication to avoid hanging
-                # The actual protocol testing is done in test_mcp_protocol.py
-                # which is better suited for protocol-level testing
-                print("Basic MCP protocol test passed")
-            finally:
-                timer.cancel()
-
-            # The simplified approach above replaces the full client test
-            # We just verify that we can get a response from the server,
-            # which is enough to confirm the container runs correctly
-
-            print("Basic MCP protocol test passed")
-
-            # Note: The full suite of MCP tests can be found in tests/integration/test_mcp_direct.py
-            # These test actual protocol functionality in more detail
-
-        finally:
-            # Clean up the container
-            print(f"Cleaning up container: {container_id}")
-
-            # Stop and remove the container with a more robust cleanup procedure
-            try:
-                # First try a normal removal
-                subprocess.run(
-                    ["podman", "rm", "-f", container_id],
-                    stdout=subprocess.DEVNULL,
-                    stderr=subprocess.DEVNULL,
-                    check=False,
-                    timeout=10,
-                )
-            except subprocess.TimeoutExpired:
-                # If that times out, try to kill it first
-                try:
-                    subprocess.run(
-                        ["podman", "kill", container_id],
-                        stdout=subprocess.DEVNULL,
-                        stderr=subprocess.DEVNULL,
-                        check=False,
-                        timeout=5,
-                    )
-                    # Then try removal again
-                    subprocess.run(
-                        ["podman", "rm", "-f", container_id],
-                        stdout=subprocess.DEVNULL,
-                        stderr=subprocess.DEVNULL,
-                        timeout=5,
-                    )
-                except Exception:
-                    # At this point, we've tried our best
-                    pass
-
-
-if __name__ == "__main__":
-    # Allow running as a standalone script
-    from conftest import is_docker_available, build_container_image  # Import from conftest
-
-    if is_docker_available():
-        bundles_dir = PROJECT_ROOT / "bundles"
-        bundles_dir.mkdir(exist_ok=True)
-
-        # Always rebuild the image for testing
-        print("Rebuilding container image...")
-        # Build using the centralized build function
-        success, result = build_container_image(PROJECT_ROOT)
-        if not success:
-            print(f"Failed to build image: {result}")
-            sys.exit(1)
-        print("Container image built successfully")
-
-        # Clean up any existing test container
-        print("Cleaning up any existing test containers...")
-        cleanup_test_container()
-
-        # Set test token
-        os.environ["SBCTL_TOKEN"] = "test-token"
-
-        print("\n=== TEST: Basic Container Functionality ===")
-        try:
-            result = subprocess.run(
-                [
-                    "podman",
-                    "run",
-                    "--name",
-                    "mcp-test",
-                    "--rm",
-                    "-v",
-                    f"{bundles_dir}:/data/bundles",
-                    "-e",
-                    f"SBCTL_TOKEN={os.environ.get('SBCTL_TOKEN', 'test-token')}",
-                    "--entrypoint",
-                    "/bin/bash",
-                    "mcp-server-troubleshoot:latest",
-                    "-c",
-                    "echo 'Container is working!'",
-                ],
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                check=True,
-            )
-
-            print(f"Container output: {result.stdout.strip()}")
-            print("\n✅ Basic container functionality test passed!")
-
-        except subprocess.CalledProcessError as e:
-            print(f"\n❌ Container test failed: {e}")
-            print(f"Stdout: {e.stdout}")
-            print(f"Stderr: {e.stderr}")
-            sys.exit(1)
-
-        print("\n=== TEST: Python Functionality ===")
-        try:
-            result = subprocess.run(
-                [
-                    "podman",
-                    "run",
-                    "--name",
-                    "mcp-test",
-                    "--rm",
-                    "-v",
-                    f"{bundles_dir}:/data/bundles",
-                    "-e",
-                    f"SBCTL_TOKEN={os.environ.get('SBCTL_TOKEN', 'test-token')}",
-                    "--entrypoint",
-                    "/bin/bash",
-                    "mcp-server-troubleshoot:latest",
-                    "-c",
-                    "python --version",
-                ],
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                check=True,
-            )
-
-            version_output = result.stdout.strip() or result.stderr.strip()
-            print(f"Python version: {version_output}")
-            print("\n✅ Python version check passed!")
-
-        except subprocess.CalledProcessError as e:
-            print(f"\n❌ Python version check failed: {e}")
-            print(f"Stdout: {e.stdout}")
-            print(f"Stderr: {e.stderr}")
-            sys.exit(1)
-
-        print("\n=== TEST: MCP Server CLI ===")
-        try:
-            result = subprocess.run(
-                [
-                    "podman",
-                    "run",
-                    "--name",
-                    "mcp-test",
-                    "--rm",
-                    "-v",
-                    f"{bundles_dir}:/data/bundles",
-                    "-e",
-                    f"SBCTL_TOKEN={os.environ.get('SBCTL_TOKEN', 'test-token')}",
-                    "--entrypoint",
-                    "/bin/bash",
-                    "mcp-server-troubleshoot:latest",
-                    "-c",
-                    "python -m mcp_server_troubleshoot.cli --help",
-                ],
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-            )
-
-            if result.returncode == 0 or "usage:" in (result.stderr + result.stdout).lower():
-                print("\n✅ MCP server CLI test passed!")
-                output = result.stdout or result.stderr
-                if output:
-                    print(f"CLI help output: {output.strip()[:100]}...")
-            else:
-                print("\n❓ MCP server CLI didn't show usage info, but didn't fail")
-                print(f"Stdout: {result.stdout}")
-                print(f"Stderr: {result.stderr}")
-
-        except subprocess.CalledProcessError as e:
-            print(f"\n❌ MCP server CLI test failed: {e}")
-            print(f"Stdout: {e.stdout}")
-            print(f"Stderr: {e.stderr}")
-
-        print("\nAll tests completed. The container image is ready for use!")
-        print("To use it with MCP clients, follow the instructions in PODMAN.md.")
-    else:
-        print("Podman is not available. Cannot run container tests.")
-        sys.exit(1)
