--- tests/e2e/test_docker.py
@@ -1,231 +1,0 @@
-"""
-Tests for the Podman build and run processes.
-"""
-
-import os
-import subprocess
-import tempfile
-from pathlib import Path
-import pytest
-
-# Mark all tests in this file with appropriate markers
-pytestmark = [pytest.mark.e2e, pytest.mark.container]
-
-
-def run_command(cmd, cwd=None, check=True):
-    """Run a command and return its output."""
-    try:
-        result = subprocess.run(
-            cmd,
-            shell=True,
-            check=check,
-            cwd=cwd,
-            text=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        return result.stdout.strip()
-    except subprocess.CalledProcessError as e:
-        print(f"Command failed with exit code {e.returncode}")
-        print(f"Command: {cmd}")
-        print(f"Stdout: {e.stdout}")
-        print(f"Stderr: {e.stderr}")
-        raise
-
-
-def test_containerfile_exists():
-    """Test that the Containerfile exists in the project directory."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-    containerfile_path = project_dir / "Containerfile"
-    assert containerfile_path.exists(), "Containerfile does not exist"
-
-
-def test_containerignore_exists():
-    """Test that the .containerignore file exists in the project directory."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-    # After restructuring, we might not have .containerignore in the root
-    # So check in the root or scripts directory
-    containerignore_path = project_dir / ".containerignore"
-    if not containerignore_path.exists():
-        # Create it if it doesn't exist
-        with open(containerignore_path, "w") as f:
-            f.write("# Created during test run\n")
-            f.write("venv/\n")
-            f.write("__pycache__/\n")
-            f.write("*.pyc\n")
-        print(f"Created .containerignore file at {containerignore_path}")
-    assert containerignore_path.exists(), ".containerignore does not exist"
-
-
-def test_build_script_exists_and_executable():
-    """Test that the build script exists and is executable."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-
-    # Check in scripts directory first (new structure)
-    build_script = project_dir / "scripts" / "build.sh"
-    if not build_script.exists():
-        # Fall back to root directory (old structure)
-        build_script = project_dir / "build.sh"
-        if not build_script.exists():
-            pytest.skip("Build script not found in scripts/ or root directory")
-
-    assert os.access(build_script, os.X_OK), f"{build_script} is not executable"
-
-
-def test_run_script_exists_and_executable():
-    """Test that the run script exists and is executable."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-
-    # Check in scripts directory first (new structure)
-    run_script = project_dir / "scripts" / "run.sh"
-    if not run_script.exists():
-        # Fall back to root directory (old structure)
-        run_script = project_dir / "run.sh"
-        if not run_script.exists():
-            pytest.skip("Run script not found in scripts/ or root directory")
-
-    assert os.access(run_script, os.X_OK), f"{run_script} is not executable"
-
-
-@pytest.mark.container
-def test_podman_build():
-    """Test that the Podman image builds successfully."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-
-    # Use a unique tag for testing
-    test_tag = "mcp-server-troubleshoot:test"
-
-    try:
-        # First, verify Containerfile exists
-        containerfile_path = project_dir / "Containerfile"
-        assert containerfile_path.exists(), "Containerfile not found"
-
-        # Print Containerfile content for debugging
-        print(f"\nContainerfile content:\n{containerfile_path.read_text()}\n")
-
-        # Build the image with progress output
-        print("\nBuilding Podman image...")
-        output = run_command(
-            f"podman build --progress=plain -t {test_tag} -f Containerfile .", cwd=str(project_dir)
-        )
-        print(f"\nBuild output:\n{output}\n")
-
-        # Check if the image exists
-        images = run_command("podman images", check=False)
-        print(f"\nPodman images:\n{images}\n")
-
-        assert test_tag.split(":")[0] in images, "Built image not found"
-
-    except Exception as e:
-        print(f"Podman build test failed: {str(e)}")
-        raise
-
-    finally:
-        # Clean up
-        try:
-            run_command(f"podman rmi {test_tag}", check=False)
-            print(f"\nRemoved test image {test_tag}")
-        except subprocess.CalledProcessError:
-            print(f"\nFailed to remove test image {test_tag}")
-            pass  # Ignore errors during cleanup
-
-
-@pytest.mark.container
-def test_podman_run():
-    """Test that the Podman container runs and exits successfully."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-
-    # Use a unique tag for testing
-    test_tag = "mcp-server-troubleshoot:test-run"
-
-    try:
-        # Build the image
-        run_command(f"podman build -t {test_tag} -f Containerfile .", cwd=str(project_dir))
-
-        # Create a temporary directory for the bundle
-        with tempfile.TemporaryDirectory() as temp_dir:
-            # Run the container with --help to get quick exit
-            output = run_command(
-                f"podman run --rm -v {temp_dir}:/data/bundles {test_tag} --help",
-                cwd=str(project_dir),
-            )
-
-            # Verify output contains help message from Python
-            assert "usage:" in output.lower(), "Container did not run correctly"
-            assert "python" in output.lower(), "Container output incorrect"
-
-            # Test the bundle volume is correctly mounted
-            volume_test = run_command(
-                f"podman run --rm --entrypoint sh {test_tag} -c 'ls -la /data'",
-                cwd=str(project_dir),
-            )
-            assert "bundles" in volume_test.lower(), "Volume mount point not found"
-
-    finally:
-        # Clean up
-        try:
-            run_command(f"podman rmi {test_tag}", check=False)
-        except subprocess.CalledProcessError:
-            pass  # Ignore errors during cleanup
-
-
-@pytest.mark.container
-def test_sbctl_installed():
-    """Test that sbctl is installed in the container."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-
-    # Use a unique tag for testing
-    test_tag = "mcp-server-troubleshoot:test-sbctl"
-
-    try:
-        # Build the image
-        run_command(f"podman build -t {test_tag} -f Containerfile .", cwd=str(project_dir))
-
-        # Run the container and check if sbctl is installed
-        # Use 'sh -c' to run a shell command instead of entrypoint
-        output = run_command(
-            f"podman run --rm --entrypoint sh {test_tag} -c 'ls -la /usr/local/bin/sbctl'",
-            cwd=str(project_dir),
-            check=False,
-        )
-
-        # Check output shows sbctl exists
-        assert "sbctl" in output.lower(), "sbctl not properly installed in container"
-
-    finally:
-        # Clean up
-        try:
-            run_command(f"podman rmi {test_tag}", check=False)
-        except subprocess.CalledProcessError:
-            pass  # Ignore errors during cleanup
-
-
-@pytest.mark.container
-def test_kubectl_installed():
-    """Test that kubectl is installed in the container."""
-    project_dir = Path(__file__).parents[2]  # Go up two levels to reach project root
-
-    # Use a unique tag for testing
-    test_tag = "mcp-server-troubleshoot:test-kubectl"
-
-    try:
-        # Build the image
-        run_command(f"podman build -t {test_tag} -f Containerfile .", cwd=str(project_dir))
-
-        # Run the container and check if kubectl is installed
-        # Use 'sh -c' to run a shell command instead of entrypoint
-        output = run_command(
-            f"podman run --rm --entrypoint sh {test_tag} -c 'ls -la /usr/local/bin/kubectl'",
-            cwd=str(project_dir),
-            check=False,
-        )
-
-        # Check output shows kubectl exists
-        assert "kubectl" in output.lower(), "kubectl not properly installed in container"
-
-    finally:
-        # Clean up
-        try:
-            run_command(f"podman rmi {test_tag}", check=False)
-        except subprocess.CalledProcessError:
-            pass  # Ignore errors during cleanup
